Use source code attached of MonoGame instead of Nu-get packages.
