+++ Shiny-lamp.sln
@@ -5,33 +5,56 @@ VisualStudioVersion = 17.5.33424.131
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Src", "Src", "{B2EADAB4-D401-4D86-A596-10D665843956}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClientApp", "Src\ClientApp\ClientApp.csproj", "{1BBF05CC-3678-4746-9398-69FB3FA24BF9}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TestTryPlay", "TestTryPlay", "{451E6D2D-29E3-46F1-9FAD-8E919C3B7A3D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TryFeature", "TestTryPlay\TryFeature\TryFeature.csproj", "{93134644-86C5-47A6-8D00-9F048E23A4E1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TryFeature", "Src\TryFeature\TryFeature.csproj", "{F4FB6A28-89C9-4C24-9A85-75D117031D1E}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Libraries", "Libraries", "{19DC66B8-E3C5-4AF7-B89C-232844CEBEAB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientApp", "Src\ClientApp\ClientApp.csproj", "{3B0C4861-81E3-443E-B61B-D708F7619024}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MonoGame.Framework.WindowsDX", "Libraries\MonoGame\MonoGame.Framework\MonoGame.Framework.WindowsDX.csproj", "{8AFB8E28-272F-4750-8487-95A10DD14388}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1BBF05CC-3678-4746-9398-69FB3FA24BF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1BBF05CC-3678-4746-9398-69FB3FA24BF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1BBF05CC-3678-4746-9398-69FB3FA24BF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1BBF05CC-3678-4746-9398-69FB3FA24BF9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{93134644-86C5-47A6-8D00-9F048E23A4E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{93134644-86C5-47A6-8D00-9F048E23A4E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{93134644-86C5-47A6-8D00-9F048E23A4E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{93134644-86C5-47A6-8D00-9F048E23A4E1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E}.Debug|x64.ActiveCfg = Debug|x64
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E}.Debug|x64.Build.0 = Debug|x64
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E}.Release|x64.ActiveCfg = Release|x64
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E}.Release|x64.Build.0 = Release|x64
+		{3B0C4861-81E3-443E-B61B-D708F7619024}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B0C4861-81E3-443E-B61B-D708F7619024}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3B0C4861-81E3-443E-B61B-D708F7619024}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3B0C4861-81E3-443E-B61B-D708F7619024}.Debug|x64.Build.0 = Debug|Any CPU
+		{3B0C4861-81E3-443E-B61B-D708F7619024}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3B0C4861-81E3-443E-B61B-D708F7619024}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3B0C4861-81E3-443E-B61B-D708F7619024}.Release|x64.ActiveCfg = Release|Any CPU
+		{3B0C4861-81E3-443E-B61B-D708F7619024}.Release|x64.Build.0 = Release|Any CPU
+		{8AFB8E28-272F-4750-8487-95A10DD14388}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8AFB8E28-272F-4750-8487-95A10DD14388}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8AFB8E28-272F-4750-8487-95A10DD14388}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8AFB8E28-272F-4750-8487-95A10DD14388}.Debug|x64.Build.0 = Debug|Any CPU
+		{8AFB8E28-272F-4750-8487-95A10DD14388}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8AFB8E28-272F-4750-8487-95A10DD14388}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8AFB8E28-272F-4750-8487-95A10DD14388}.Release|x64.ActiveCfg = Release|Any CPU
+		{8AFB8E28-272F-4750-8487-95A10DD14388}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{1BBF05CC-3678-4746-9398-69FB3FA24BF9} = {B2EADAB4-D401-4D86-A596-10D665843956}
-		{93134644-86C5-47A6-8D00-9F048E23A4E1} = {451E6D2D-29E3-46F1-9FAD-8E919C3B7A3D}
+		{F4FB6A28-89C9-4C24-9A85-75D117031D1E} = {451E6D2D-29E3-46F1-9FAD-8E919C3B7A3D}
+		{3B0C4861-81E3-443E-B61B-D708F7619024} = {B2EADAB4-D401-4D86-A596-10D665843956}
+		{8AFB8E28-272F-4750-8487-95A10DD14388} = {19DC66B8-E3C5-4AF7-B89C-232844CEBEAB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A94D3FCF-2A6F-432E-BC57-02F54D59F1EB}
