--- Src/ClientApp/Characters/JobCharacter.cs
@@ -1,10 +1,0 @@
-﻿using ClientApp.Characters.Abstracts;
-using System.Collections.Generic;
-
-namespace ClientApp.Characters
-{
-    internal class JobCharacter : Character
-    {
-        
-    }
-}
