--- Src/ClientApp/Characters/NPCCharacter.cs
@@ -1,8 +1,0 @@
-﻿using ClientApp.Characters.Abstracts;
-
-namespace ClientApp.Characters
-{
-    internal class NPCCharacter : Character
-    {
-    }
-}
