+++ Src/ClientApp/Program.cs
@@ -1,13 +1,3 @@
-﻿using ClientApp.ClientSource;
-
-namespace ClientApp
-{
-    internal class Program
-    {
-        static void Main(string[] args)
-        {
-            using var game = new GameMain();
-            game.Run();
-        }
-    }
-}
+﻿
+using var game = new ClientApp.Game1();
+game.Run();
