+++ Src/TryFeature/Characters/Abstracts/Character.cs
