+++ Src/TryFeature/Characters/JobCharacter.cs
